Add tests for all services

+++ RetroWars.Services.Data/PollService.cs
@@ -7,7 +7,6 @@ using Retrowars.Data.Repository;
 using Contracts;
 using Web.ViewModels.Poll;
 
-
 public class PollService : IPollService
 {
     private readonly IRepository<Poll> pollRepository;
@@ -101,19 +100,29 @@ public class PollService : IPollService
         return model;
     }
 
-    public async Task CreatePollAsync(PollFormModel formModel)
+    public async Task<bool> CreatePollAsync(PollFormModel formModel)
     {
-        Poll newPoll = new Poll()
+        try
         {
-            FirstGameId = Guid.Parse(formModel.FirstGameId),
-            SecondGameId = Guid.Parse(formModel.SecondGameId),
-            IsActive = true,
-            VotesForFirst = 0,
-            VotesForSecond = 0,
-        };
+            Poll newPoll = new Poll()
+            {
+                FirstGameId = Guid.Parse(formModel.FirstGameId),
+                SecondGameId = Guid.Parse(formModel.SecondGameId),
+                IsActive = true,
+                VotesForFirst = 0,
+                VotesForSecond = 0,
+            };
 
-        await this.pollRepository.AddAsync(newPoll);
-        await this.pollRepository.SaveAsync();
+            await this.pollRepository.AddAsync(newPoll);
+            await this.pollRepository.SaveAsync();
+
+            return true;
+        }
+        catch
+        {
+
+            return false;
+        }
     }
 
 
@@ -173,22 +182,35 @@ public class PollService : IPollService
 
     }
 
-    public async Task MarkUserAsVoted(string id, string userId)
+    public async Task<bool> MarkUserAsVoted(string id, string userId)
     {
-       Poll? poll  = await this.pollRepository.GetOneAsync(id, false);
-       ApplicationUser? user = await this.applicationUserRepository.GetOneAsync(userId, false);
-       if (poll is null || user is null)
-       {
-           throw new ArgumentException("Invalid poll or user id.");
-       }
-       poll.Voters.Add(user);
-       await this.pollRepository.SaveAsync();
+        try
+        {
+            Poll? poll = await this.pollRepository.GetOneAsync(id, false);
+            ApplicationUser? user = await this.applicationUserRepository.GetOneAsync(userId, false);
+
+
+            if (poll is null || user is null)
+            {
+                throw new ArgumentException("Invalid poll or user id.");
+            }
+            poll.Voters.Add(user);
+            await this.pollRepository.SaveAsync();
+            return true;
+        }
+        catch (Exception)
+        {
+
+            return false;
+        }
     }
 
     public double[] GetResults(int votesForFirst, int votesForSecond)
     {
         double[] result = new double[2];
-
+        if(votesForFirst  == 0 && votesForSecond == 0) {
+            return result;
+        }
         result[0] = (double)votesForFirst/(votesForFirst+votesForSecond);
         result[1] = (double)votesForSecond/(votesForSecond+votesForFirst);
 
@@ -197,30 +219,47 @@ public class PollService : IPollService
 
 
     
-    public async Task DeactivateAPoll(string id)
+    public async Task<bool> DeactivateAPoll(string id)
     {
-      Poll? poll = await  this.pollRepository.GetOneAsync(id, false);
+        try
+        {
+            Poll? poll = await this.pollRepository.GetOneAsync(id, false);
 
-      if (poll is null)
-      {
-          throw new ArgumentException("Invalid id");
-      }
+            if (poll is null)
+            {
+                throw new ArgumentException("Invalid id");
+            }
 
-      poll.IsActive = false;
-      await this.pollRepository.SaveAsync();
+            poll.IsActive = false;
+            await this.pollRepository.SaveAsync();
+            return true;
+        }
+        catch 
+        {
+
+            return false;
+        }
     }
       
-    public async Task ActivateAPoll(string id)
+    public async Task<bool> ActivateAPoll(string id)
     {
-        Poll? poll = await this.pollRepository.GetOneAsync(id, false);
-
-        if (poll is null)
+        try
         {
-            throw new ArgumentException("Invalid id");
-        }
+            Poll? poll = await this.pollRepository.GetOneAsync(id, false);
 
-        poll.IsActive = true;
-        await this.pollRepository.SaveAsync();
+            if (poll is null)
+            {
+                throw new ArgumentException("Invalid id");
+            }
+
+            poll.IsActive = true;
+            await this.pollRepository.SaveAsync();
+            return true;
+        }
+        catch 
+        {
+            return false;
+        }
     }
 
     public async Task<IEnumerable<PollAdminViewModel>> GetAllPollAdminViewModels()
@@ -230,7 +269,7 @@ public class PollService : IPollService
         IEnumerable<PollAdminViewModel> allPollViewModels = allPolls.Select(p => new PollAdminViewModel()
         {
             Id = p.Id,
-            SecondGameId = p.SecondGameId,
+           
             FirstGameName = p.FirstGame.Name,
             SecondGameName = p.SecondGame.Name,
             FirstGamePlatform = p.FirstGame.Platform.Name,
@@ -246,10 +285,26 @@ public class PollService : IPollService
         return allPollViewModels;
     }
 
-    public async Task DeletePoll(string id)
+    public async Task<bool> DeletePoll(string id)
     {
-        await this.pollRepository.DeleteOneAsync(Guid.Parse(id));
-        await this.pollRepository.SaveAsync();
+        try
+        {
+            Poll? toDelete = await this.pollRepository.GetOneAsync(id,false);
+            if(toDelete is null)
+            {
+                throw new ArgumentException("Invalid id.");
+            }
+            await this.pollRepository.DeleteOneAsync(Guid.Parse(id));
+
+            
+            await this.pollRepository.SaveAsync();
+            return true;
+        }
+        catch 
+        {
+
+          return false;
+        }
     }
 }
